requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain_community
PyPDF2
faiss-cpu
langchain_google_genai
langchain_ollama

bs4
pypdf
supabase
fastapi 
uvicorn
pydantic

langchain_groq
langchain_core

langserve
sse_starlette